requirements.txt
asgiref==3.5.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.1.0
idna==3.3
mysqlclient==2.1.0
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pytz==2022.1
qiniu==7.7.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
